# Conflicts:
#	learn-cloud-common/src/main/resources/bootstrap.yml
#	learn-shop-admin-system/src/main/java/com/billow/system/api/ActProcDefApi.java
#	learn-shop-public-auth/src/main/java/com/billow/auth/security/config/WebSecurityConfig.java
#	learn-shop-ui-admin/package-lock.json

+++ learn-shop-admin-system/src/main/java/com/billow/system/service/MenuService.java
@@ -70,7 +70,7 @@ public interface MenuService extends IService<MenuPo> {
      * @author LiuYongTao
      * @date 2019/7/24 14:10
      */
-    Integer countMenuCodeByMenuCode(String menuCode);
+    Long countMenuCodeByMenuCode(String menuCode);
 
     /**
      * 查询出指定节点的所有直接父级菜单id
